requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
PyMuPDF==1.23.22
langchain==0.1.20
langchain-core==0.1.51
langchain-community==0.0.34
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
langchain-chroma==0.1.0
chromadb==0.4.24
google-generativeai==0.4.1
